split headers into source, header
