+++ .gitignore
@@ -1,0 +1,4 @@
+*.o
+ceec
+out.asm
+compile_commands.json
